requirements.txt
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
django-filter==23.5
dj-database-url==2.1.*
psycopg2-binary==2.9.*
python-decouple==3.8
Pillow==10.*
cloudinary==1.36.*
django-cloudinary-storage==0.3.*
razorpay==1.4.*
celery==5.3.*
redis==5.0.*
django-celery-beat==2.5.*
django-celery-results==2.5.*
gunicorn==21.*
whitenoise==6.6.*
setuptools>=70
